feat: use rules for test filters
