+++ run-tests.js
@@ -13,68 +13,16 @@ const { createNextInstall } = require('./test/lib/create-next-install')
 const glob = promisify(_glob)
 const exec = promisify(execOrig)
 const core = require('@actions/core')
+const { getTestFilter } = require('./test/get-test-filter')
 
 function escapeRegexp(str) {
   return str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
 }
 
-/**
- * @typedef {{ file: string, excludedCases: string[] }} TestFile
- */
-
 const GROUP = process.env.CI ? '##[group]' : ''
 const ENDGROUP = process.env.CI ? '##[endgroup]' : ''
 
-function getTestFilters() {
-  const manifest = process.env.NEXT_EXTERNAL_TESTS_FILTERS
-    ? require(path.resolve(process.env.NEXT_EXTERNAL_TESTS_FILTERS))
-    : null
-  if (!manifest) return null
-
-  // For the legacy manifest without a version, we assume it's a complete list
-  // of all the tests.
-  if (!manifest.version || typeof manifest.version !== 'number') {
-    return (tests) =>
-      tests
-        .filter((test) => {
-          const info = manifest[test.file]
-          return info && info.passed.length > 0 && !info.runtimeError
-        })
-        .map((test) => {
-          const info = manifest[test.file]
-          // Exclude failing and flakey tests, newly added tests are automatically included
-          if (info.failed.length > 0 || info.flakey.length > 0) {
-            test.excludedCases = info.failed.concat(info.flakey)
-          }
-          return test
-        })
-  }
-
-  // The new manifest version 2 only contains the list of tests that should
-  // be run, with exclusions added.
-  if (manifest.version === 2) {
-    return (tests) =>
-      tests
-        // Only include tests that are in the manifest.
-        .filter((test) => {
-          const info = manifest.suites[test.file]
-          return info && !info.runtimeError
-        })
-        .map((test) => {
-          const { failed = [], flakey = [] } = manifest.suites[test.file]
-
-          // Exclude failing and flakey tests, newly added tests are automatically included
-          if (failed.length > 0 || flakey.length > 0) {
-            test.excludedCases = failed.concat(flakey)
-          }
-          return test
-        })
-  }
-
-  throw new Error(`Unknown manifest version: ${manifest.version}`)
-}
-
-const externalTestsFilter = getTestFilters()
+const externalTestsFilter = getTestFilter()
 
 const timings = []
 const DEFAULT_NUM_RETRIES = os.platform() === 'win32' ? 2 : 1
